# Conflicts:
#	hr-cpp/hr-c7-datasetting/dao/archive/ArchiveDAO.h
#	hr-cpp/hr-c7-datasetting/domain/do/archive/ArchiveDo.h
#	hr-cpp/hr-c7-datasetting/domain/dto/archive/ArchiveDTO.h
#	hr-cpp/hr-c7-datasetting/hr-c7-datasetting.vcxproj
#	hr-cpp/hr-c7-datasetting/hr-c7-datasetting.vcxproj.filters
#	hr-cpp/hr-c7-datasetting/main.cpp
#	hr-cpp/hr-c7-datasetting/service/archive/ArchiveService.cpp
#	hr-cpp/hr-c7-datasetting/service/archive/ArchiveService.h

+++ hr-cpp/hr-c7-datasetting/domain/dto/archive/ArchiveDTO.h
@@ -11,7 +11,7 @@ class ArchiveDTO : public oatpp::DTO
 {
 	DTO_INIT(ArchiveDTO, DTO);
 	// ÅÅÐòºÅ
-	DTO_FIELD(UInt32, sortid);
+	DTO_FIELD(UInt64, sortid);
 	DTO_FIELD_INFO(sortid) {
 		info->description = ZH_WORDS_GETTER("archive.field.sortid");
 	}
